@@ -5,24 +5,34 @@
   "main": "webpack.config.js",
   "dependencies": {
     "@types/createjs": "0.0.29",
+    "@types/jquery": "^3.3.6",
     "@types/js-cookie": "^2.1.0",
     "@types/js-yaml": "^3.11.2",
     "@types/node-uuid": "0.0.28",
-    "@types/socket.io": "^1.4.31",
+    "@types/socket.io": "^1.4.36",
     "@types/socket.io-client": "^1.4.32",
-    "express": "^4.16.2",
+    "express": "^4.16.3",
+    "jquery": "^3.3.1",
     "js-cookie": "^2.2.0",
     "js-yaml": "^3.12.0",
     "node-uuid": "^1.4.8",
-    "router": "^1.3.2",
-    "socket.io": "^2.0.4",
-    "ts-loader": "^3.2.0",
-    "typescript": "^2.8.3"
+    "router": "^1.3.3",
+    "socket.io": "^2.1.1",
+    "ts-loader": "^4",
+    "typescript": "^3"
   },
   "devDependencies": {
     "@types/jasmine": "^2.8.8",
-    "jasmine": "^3.1.0",
-    "webpack": "^3.12.0"
+    "gulp": "^3.9.1",
+    "gulp-concat": "^2.6.1",
+    "gulp-plumber": "^1.2.0",
+    "gulp-typescript": "^5.0.0-alpha.3",
+    "jasmine": "^3.2.0",
+    "require-dir": "^1.0.0",
+    "run-sequence": "^2.2.1",
+    "webpack": "^4.14.0",
+    "webpack-cli": "^3.1.0",
+    "webpack-stream": "^5.1.1"
   },
   "scripts": {
     "start": "node dist/Server/main.js",
