@@ -1,105 +1,125 @@
 import { SocketBinder } from "./socketBinder";
-import { GamePlayerState } from "../Share/gamePlayerState";
+import { ResponseGamePlayerState } from "../Share/responseGamePlayerState";
 import { PlayerData } from "./playerData";
 import { SocketBinderList } from "./socketBinderList";
 import { DiceNumber } from "../Share/diceNumber";
-import { ResourceIndex, GenerateResourceYamlData } from "../Share/Yaml/resourceYamlData";
-import { yamlGet } from "../Share/Yaml/yamlGet";
-import { GenerateActionCardYamlData } from "../Share/Yaml/actionCardYamlDataGen";
-import { BuildActionIndex, ActionCardYamlData } from "../Share/Yaml/actionCardYamlData";
+import { ResourceName } from "../Share/Yaml/resourceYamlData";
+import { GamePlayerCondition } from "../Share/gamePlayerCondition";
+import { ActionCardName, ActionCardYamlData } from "../Share/Yaml/actionCardYamlData";
+import { GamePlayerState } from "./gamePlayerState";
+import { StartStatusYamlData } from "../Share/Yaml/startStatusYamlData";
 
 export class GamePlayer {
     private playerId: number;
     private uuid: string;
-    private state: SocketBinder<GamePlayerState>;
-    private resourceList: SocketBinderList<ResourceIndex>;
-    private buildActionList: SocketBinderList<BuildActionIndex>;
+    private state: GamePlayerState;
+    private resourceList: SocketBinderList<ResourceName>;
+    private buildActionList: SocketBinderList<ActionCardName>;
     private diceList: SocketBinder<DiceNumber[]>;
-    private actionCardList: SocketBinderList<ActionCardYamlData | null>;
+    private actionCardList: SocketBinderList<string | null>;
+    private playerCond: SocketBinder<GamePlayerCondition>;
+    private isGameMaster: boolean = false;
 
     get Uuid() { return this.uuid; }
     get PlayerId() { return this.playerId; }
+    get IsGameMaster() { return this.isGameMaster; }
+    set IsGameMaster(x) { this.isGameMaster = x; }
+
+    setAICard(ai: StartStatusYamlData) { this.state.setAICard(ai); }
+
+    setMyTurn() {
+        this.playerCond.Value = GamePlayerCondition.MyTurn;
+    }
+
+    setWait() {
+        this.playerCond.Value = GamePlayerCondition.Wait;
+    }
 
     constructor(
         playerData: PlayerData,
         playerId: number,
-        state: SocketBinder<GamePlayerState>,
-        resourceList: SocketBinderList<ResourceIndex>,
-        buildActionList: SocketBinderList<BuildActionIndex>,
+        state: SocketBinder<ResponseGamePlayerState>,
+        resourceList: SocketBinderList<ResourceName>,
+        buildActionList: SocketBinderList<ActionCardName>,
         diceList: SocketBinder<DiceNumber[]>,
-        actionCardList: SocketBinderList<ActionCardYamlData | null>
+        actionCardList: SocketBinderList<string | null>,
+        playerCond: SocketBinder<GamePlayerCondition>
     ) {
         this.diceList = diceList;
         this.diceList.Value = [0, 1, 2];
         this.playerId = playerId;
         this.uuid = playerData.getUuid();
-        this.state = state;
+        this.state = new GamePlayerState(state, playerData.getName());
         this.resourceList = resourceList;
         this.buildActionList = buildActionList;
         this.actionCardList = actionCardList;
-        this.state.Value = {
-            playerName: playerData.getName(),
-            negative: 0, positive: 0,
-            uncertainty: 0, resource: 0,
-            activityRange: 0, speed: 0
-        };
-        const buildAction = GenerateActionCardYamlData(yamlGet("./Resource/yamls/actionCard.yaml"), true);
         this.buildActionList.Value = [
-            buildAction["採掘施設"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["教会"].index,
-            buildAction["核融合炉"].index,
-            buildAction["ロボット工場"].index,
+            "採掘施設",
+            "治療施設",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "教会",
+            "核融合炉",
+            "ロボット工場",
         ];
-        const resourceAction = GenerateResourceYamlData(yamlGet("./Resource/yamls/resource.yaml"));
         this.resourceList.Value = [
-            resourceAction["人間"].index,
-            resourceAction["人間"].index,
-            resourceAction["人間"].index,
-            resourceAction["人間"].index,
-            resourceAction["人間"].index,
-            resourceAction["聖書"].index,
-            resourceAction["聖書"].index,
-            resourceAction["CPU"].index,
-            resourceAction["CPU"].index,
-            resourceAction["CPU"].index,
-            resourceAction["CPU"].index,
-            resourceAction["CPU"].index,
-            resourceAction["CPU"].index,
-            resourceAction["CPU"].index,
-            resourceAction["拡張人間"].index,
-            resourceAction["拡張人間"].index,
-            resourceAction["拡張人間"].index,
-            resourceAction["拡張人間"].index,
-            resourceAction["拡張人間"].index,
-            resourceAction["拡張人間"].index,
-            resourceAction["拡張人間"].index,
+            "人間",
+            "人間",
+            "人間",
+            "人間",
+            "人間",
+            "聖書",
+            "聖書",
+            "CPU",
+            "CPU",
+            "CPU",
+            "CPU",
+            "CPU",
+            "CPU",
+            "CPU",
+            "拡張人間",
+            "拡張人間",
+            "拡張人間",
+            "拡張人間",
+            "拡張人間",
+            "拡張人間",
+            "拡張人間",
+            "テラフォーミング",
         ];
-        const actionCard = GenerateActionCardYamlData(yamlGet("./Resource/yamls/actionCard.yaml"), false);
-        actionCardList.Value = [null, null, actionCard["神の杖"], null, null]
-        actionCardList.setAt(0, actionCard["意識操作のテスト"])
+        actionCardList.Value = [null, null, null, null, null];
+        this.playerCond = playerCond;
+        playerCond.Value = GamePlayerCondition.Start;
 
     }
 
     //ユーザーにsocketBinderの値を送信する
     sendToSocket(socket: SocketIO.Socket) {
-        this.state.updateAt(socket);
+        this.state.sendToSocket(socket);
         this.resourceList.updateAt(socket);
         this.buildActionList.updateAt(socket);
         this.diceList.updateAt(socket);
-        this.actionCardList.updateAt(socket);
+    }
+
+    addSocket(socket: SocketIO.Socket) {
+        this.actionCardList.addSocket(socket);
+        this.playerCond.addSocket(socket);
+    }
+    drawActionCard(card: ActionCardYamlData) {
+        const index = this.actionCardList.Value.findIndex(x => x == null);
+        if (index == -1)
+            throw "手札がいっぱいです";
+        this.actionCardList.setAt(index, card.name);
     }
 }
